Create Document superclass for Item and Layout
