+++ lib/nanoc/base/entities.rb
@@ -1,2 +1,3 @@
+require_relative 'entities/document'
 require_relative 'entities/identifier'
 require_relative 'entities/pattern'
